Final Graded assessment of Automation using Python
Final Graded assessment of Automation using Python.. Cleared the assessment on 29-Aug-2021. This assessment took two attempts as I was not sure about SMTP server to send emails from in the test server. Beat my head up but found it out to be as stupid as localhost. I should have checked the source code of previous graded assessment where the code for similar action is provided.